[bug #166660718] Fixing backend LFA suggestion
