--- .env
@@ -1,1 +1,0 @@
-SECRETKEY='automart123@carlos?andela(developer)'
