+++ .gitignore
@@ -1,5 +1,5 @@
 coverage
 .nyc_output
-
+.env
 node_modules/
 package-lock.json
