+++ server/helper/cloudinaryConfig.js
@@ -1,9 +1,12 @@
 import cloudinary from 'cloudinary';
+import dotenv from 'dotenv';
+
+dotenv.config();
 
 const cloudinaryConfig = cloudinary.config({
-  cloud_name: "dnuqiengg",
-  api_key: "929645198658237",
-  api_secret: "5nQhn1XWTRuwfrfwxUZ1rj-Q_uw"
+  cloud_name: process.env.cloud_name,
+  api_key: process.env.api_key,
+  api_secret: process.env.api_secret
 });
 
 export default cloudinaryConfig;
