+++ server/helper/validation.js
@@ -20,7 +20,7 @@ const validateCar = Joi.object({
 	manufacture: Joi.string().alphanum().min(3).required(),
 	model: Joi.string().min(3).required(),
 	price: Joi.number().precision(4).positive().min(2).required(),
-	state: Joi.string().min(3).required(),
+	state: Joi.string().valid('new', 'used').required(),
 	body_type: Joi.string().min(3).required(),
 	description: Joi.string().max(150).required(),
 });
