+++ server/models/cars.js
@@ -1,46 +1,4 @@
 const cars = [];
 
-const car1 = {
-	"id": 1,
-    "owner": 2,
-    "created_on": "2019-5-26 15:51:21",
-    "state": "new",
-    "status": "available",
-    "price": "1200.23",
-    "manufacture": "toyota",
-    "model": "seria A",
-    "body_type": "sport car",
-    "description": "hsdhdbhabhsds hey there mamo",
-    "image": "https://res.cloudinary.com/dnuqiengg/image/upload/v1558882677/jbpf3tm3gvsm11ile2tn.jpg"
-};
 
-const car2 = {
-    "id": 2,
-    "owner": 3,
-    "created_on": "2019-5-26 16:2:13",
-    "state": "new",
-    "status": "available",
-    "price": "1200.23",
-    "manufacture": "BMW",
-    "model": "seria A",
-    "body_type": "sport car",
-    "description": "hsdhdbhabhsds hey there mamo",
-    "image": "https://res.cloudinary.com/dnuqiengg/image/upload/v1558882934/dlqpj74qdatssyjw7mb4.jpg"
-};
-
-const car3 = {
-	"id": 3,
-    "owner": 2,
-    "created_on": "2019-5-26 16:2:13",
-    "state": "new",
-    "status": "sold",
-    "price": "1200.23",
-    "manufacture": "BMW",
-    "model": "seria A",
-    "body_type": "sport car",
-    "description": "hsdhdbhabhsds hey there mamo",
-    "image": "https://res.cloudinary.com/dnuqiengg/image/upload/v1558882934/dlqpj74qdatssyjw7mb4.jpg"	
-};
-
-cars.push(car1,car2,car3);
 export default cars;
