+++ server/routes/cars/cars.js
@@ -1,6 +1,7 @@
 import express from 'express';
 import Car from '../../controllers/car';
 import auth from '../../middlewares/checkAuth';
+import admin from '../../middlewares/admin';
 import multipart from 'connect-multiparty';
 const multipartMiddleware = multipart();
 
@@ -23,11 +24,15 @@ router.patch('/:id/status', auth, car.updateStatus);
 // view a specific car
 router.get('/:id', car.fetchId);
 
-// view a specific car
+// view a car with option like status or price range
 router.get('/', car.fetch);
 
+// admin view all cars
+
+// router.get('/all', car.fetchAll) 
+
 // Delete a specific car
-router.delete('/:id', auth, car.delete);
+router.delete('/:id', [auth, admin], car.delete);
 
 // reporting a car
 router.post('/flag', auth, car.report);
